AI/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "google-genai>=1.60.0",
    "supabase>=2.25.1",
]
